BUG checkPointByte should be static_cast to the int8_t

+++ src/CAta_Ext_DST_Log.cpp
@@ -200,7 +200,7 @@ eReturnValues CAta_Ext_DST_Log::parse_Ext_Self_Test_Log(JSONNODE *masterData)
         StatusByte = pData[DSTIndex + 1];
         timeStamp = M_BytesTo2ByteValue(pData[DSTIndex + 3], pData[DSTIndex + 2]);
         compTime = M_BytesTo2ByteValue(pData[DSTIndex + 13], pData[DSTIndex + 12]);
-        checkPointByte = pData[DSTIndex + 4];
+        checkPointByte = static_cast<int8_t>(pData[DSTIndex + 4]);
         LBA = M_BytesTo8ByteValue(0,0, pData[DSTIndex + 10], pData[DSTIndex + 9], pData[DSTIndex + 8], pData[DSTIndex + 7], pData[DSTIndex + 6], pData[DSTIndex + 5]);
         JSONNODE *runInfo = json_new(JSON_NODE);
